🛠️refactor: useGalleryMoving 로직 분리
