+++ src/app/(Product)/wedding/invite/[id]/_components/InviteGallery/components/GalleryModal/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useGalleryMoving } from './useGalleryMoving/useGalleryMoving';
