+++ src/app/(Product)/wedding/invite/[id]/_components/InviteGallery/components/GalleryModal/hooks/useGalleryMoving/useGalleryMoving.type.ts
@@ -1,0 +1,4 @@
+export interface UseGalleryMovingProps {
+  viewNumber: number;
+  maxIndex: number;
+}
